pdf_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
fpdf==1.7.2
boto3==1.34.0
PyJWT==2.8.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0 